refactor: restructure MobileNav
